fix(net): parse remote player level as short

+++ src/client/Character/OtherChar.cpp
@@ -21,7 +21,7 @@
 
 namespace jrc
 {
-    OtherChar::OtherChar(int32_t id, const CharLook& lk, uint8_t lvl,
+    OtherChar::OtherChar(int32_t id, const CharLook& lk, uint16_t lvl,
         int16_t jb, const std::string& nm, int8_t st, Point<int16_t> pos) : Char(id, lk, nm) {
 
         level = lvl;
